Moves files into logic packages.
