+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme" >
-        <service android:name=".LocationProviderService" />
+        <service android:name=".utils.LocationProviderService" />
         <activity
             android:name=".RouterActivity"
             android:label="@string/app_name"
