+++ app/src/main/java/xyz/whereuat/whereuat/utils/LocationProviderService.java
@@ -1,4 +1,4 @@
-package xyz.whereuat.whereuat;
+package xyz.whereuat.whereuat.utils;
 
 import android.app.Service;
 import android.content.Intent;
