+++ app/src/main/res/layout/contact_card.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<xyz.whereuat.whereuat.ContactCard xmlns:android="http://schemas.android.com/apk/res/android"
+<xyz.whereuat.whereuat.ui.views.ContactCard xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/container_flipper"
     android:layout_width="match_parent"
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:padding="@dimen/spacing_margin" >
-        <xyz.whereuat.whereuat.AutoShareStar
+        <xyz.whereuat.whereuat.ui.views.AutoShareStar
             android:id="@+id/auto_share_status"
             android:layout_alignParentRight="true"
             android:layout_alignParentTop="true"
@@ -18,7 +18,7 @@
             android:layout_width="@dimen/auto_share_star_width"
             app:draw_color="@color/offWhite"
             app:is_clickable="false" />
-        <xyz.whereuat.whereuat.LatoTextView
+        <xyz.whereuat.whereuat.ui.views.LatoTextView
             android:id="@+id/front_view_initials"
             app:lato_font="semibold"
             android:layout_centerInParent="true"
@@ -26,7 +26,7 @@
             android:layout_width="wrap_content"
             android:textColor="@color/offWhite"
             android:textSize="@dimen/front_initials_font_size" />
-        <xyz.whereuat.whereuat.LatoTextView
+        <xyz.whereuat.whereuat.ui.views.LatoTextView
             android:id="@+id/front_view_fullname"
             android:ellipsize="end"
             app:lato_font="regular"
@@ -45,7 +45,7 @@
         android:layout_height="fill_parent"
         android:layout_width="fill_parent"
         android:padding="@dimen/spacing_margin">
-        <xyz.whereuat.whereuat.LatoTextView
+        <xyz.whereuat.whereuat.ui.views.LatoTextView
             android:id="@+id/back_view_fullname"
             app:lato_font="bold"
             android:layout_alignParentTop="true"
@@ -62,7 +62,7 @@
             android:layout_height="wrap_content"
             android:orientation="horizontal"
             android:layout_centerInParent="true">
-            <xyz.whereuat.whereuat.LatoTextView
+            <xyz.whereuat.whereuat.ui.views.LatoTextView
                 android:id="@+id/auto_share_text"
                 app:lato_font="regular"
                 android:layout_width="wrap_content"
@@ -70,7 +70,7 @@
                 android:text="Auto Share?"
                 android:textColor="@color/darkGray"
                 android:textSize="18sp" />
-            <xyz.whereuat.whereuat.AutoShareStar
+            <xyz.whereuat.whereuat.ui.views.AutoShareStar
                 android:id="@+id/auto_share_button"
                 android:layout_width="@dimen/auto_share_star_width"
                 android:layout_height="@dimen/auto_share_star_height"
@@ -88,7 +88,7 @@
             android:gravity="center"
             android:layout_alignParentBottom="true"
             android:layout_centerInParent="true">
-            <xyz.whereuat.whereuat.LatoTextView
+            <xyz.whereuat.whereuat.ui.views.LatoTextView
                 app:lato_font="regular"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -96,7 +96,7 @@
                 android:textColor="@color/darkGray"
                 android:textSize="18sp"
                 android:textAlignment="center" />
-            <xyz.whereuat.whereuat.LatoTextView
+            <xyz.whereuat.whereuat.ui.views.LatoTextView
                 android:id="@+id/requested_text"
                 app:lato_font="regular"
                 android:layout_width="wrap_content"
@@ -107,4 +107,4 @@
                 android:textAlignment="center" />
         </LinearLayout>
     </RelativeLayout>
-</xyz.whereuat.whereuat.ContactCard>
+</xyz.whereuat.whereuat.ui.views.ContactCard>
